// packages/draw/core/src/types.ts
export type StringValue = string | Ref<string>;

type IdValue = string;
export interface Identifiable {
    id: IdValue
}
export interface BlockElement extends Identifiable {
    label: StringValue
    parentId?: IdValue
    enter: EnterSetting
    debug?: DebugSetting
}

export type ExcludeProp<T, U> = {
    [Property in keyof T as Exclude<Property, U>]: T[Property]
};
export interface LineBlockElement extends ExcludeProp<LineBlockElementDescription, 'label'>, Identifiable {
    label: StringValue
}

export interface ParentBlockElement extends BlockElement {
    children: Block[]
}

export type ContainerBlock = BlockElement | ParentBlockElement;
export type Block = ContainerBlock | LineBlockElement;
// ...
export type BlocksStyle = Record<string, BlockStyle>;
export interface Style extends Identifiable {
    positioning: BlockPositioning[]
    blocks?: BlocksStyle
    debug?: GlobalDebugSetting
    css?: StyleCss
}
// ...
export interface BlockPositioning {
    blockId: IdValue
    position: Positioning
}

export interface ArchitectureData {
    blocks: Block[]
    style?: Style | null

}
export interface Architecture extends RefsType<ArchitectureData> {
    name: string
    phases: PhaseId[]
    currentPhase: Ref<PhaseId> | null
    next: () => PhaseId
    back: () => PhaseId
    debugSelect: (select: SelectedProperties) => void
    toPhase: (phase: number | null) => void
    debugState?: Ref<DebugStateContainer | undefined>
}
export type RefsType<T> = {
    [P in keyof T]?: Ref<T[P]>;
};
export interface TransitionalArchitecture extends ArchitectureDescription {
}

export type PhaseId = number | null | undefined;
// ...
export interface PhaseIndex {
    getNextPhaseById(current: PhaseId): PhaseIndexItem | undefined
    getPhaseById(current: PhaseId): PhaseIndexItem | undefined
    phases: PhaseId[]
    getPhaseIndex(phase?: number | string | null): number
    findPhaseIdBy: (func: (phaseIndex: PhaseIndexItem) => boolean) => PhaseId
}
export interface PhaseIndexItem {
    phaseId: number
    hasNext: boolean
    actions: PhaseIndexItemAction[]
}
// ...
export interface PropertiesConfig {
    name: string
    index: number
}
export interface SelectedProperties {
    properties: PropertiesConfig[]
    file: string
}
// ...
export interface DebugStateContainer {
    selectedBlocks?: string[] | null
    lines?: LineDebugAction[]
}
// ...
export interface DebugSelectContext {
    style: Ref<Style>
    blocks: Ref<Block[]>
    phaseIndex: PhaseIndex
    transformSetting: TransformSetting
}
export interface TransformationResult {
    architecture: Ref<Architecture>
    config: Ref<ArchitectureDescription>
}